+++ xray/__init__.py
@@ -4,7 +4,7 @@ from .core.dataset import Dataset
 from .core.dataarray import DataArray
 from .core.options import set_options
 
-from .backends.api import open_dataset, open_mfdataset
+from .backends.api import open_dataset, open_mfdataset, save_mfdataset
 from .conventions import decode_cf
 
 from .version import version as __version__